So, the value I had though was the object id is not.  Fix.
In the third segment, it appeared that there was an object id, related back
to the object id given in a structure.  However, this is not true at all, as
MATLAB will randomly decide what values go here.  However, the index into this
segment seems to be the actual object id.  Thus switch to using that value.
The meaning of the other field is left for someone else to divine, it isn't
important here.