Change MLObject to use a Map<String,MLArray> to store properties.
With the change to have MLObject properly store arrays of objects, and due
to the fact MLObject needs to handle the handle class, using MLStructure as
the data storage and array handler is about to get very messy.  Thus change
to the obvious data structure to manage everything.  The instance of a Map<>
is a specific instance of an MCOS.  That way handles are properly represented.